CHG: modify the poetry installation; add makefile; remove the cache folders

+++ envs/conda/utils.sh
@@ -40,8 +40,7 @@ initialize_conda_env() {
 
   conda create -c conda-forge -n "${CONDA_ENV}" "python=${PYTHON_VERSION}" -y
   source activate "${CONDA_ENV}"
-  conda install pipx -y
-  pipx install poetry --force
+  pip install --no-cache-dir "poetry==${POETRY_VERSION}"
   conda deactivate
 }
 
@@ -78,7 +77,12 @@ install_python_package() {
   echo -e "${FG_YELLOW}Installing python package${FG_RESET}"
   poetry install
   poetry build
-  pip install dist/*.tar.gz
-  rm -r dist
+  if [ -d "${PWD}"/dist/ ]; then
+    pip install dist/*.tar.gz
+    rm -r dist
+  else
+    echo -e "${FG_RED}Failed to install python package${FG_RESET}"
+  fi
+
   popd
 }
